views relocation checks, p.0

+++ nd4j-backends/nd4j-backend-impls/nd4j-cuda-7.5/src/test/java/jcuda/jcublas/ops/AveragingTests.java
@@ -138,7 +138,7 @@ public class AveragingTests {
             INDArray array = pairs.get(i).getKey();
             AllocationPoint point = allocator.getAllocationPoint(array.data());
 
-            if (point.getDeviceId().intValue() != 0 )
+            if (point.getDeviceId() != 0 )
                 hasNonZero.set(true);
 
             arrays.add(array);
